+++ workersparadise/src/main/resources/templates/fragments/_navbar.html
@@ -7,5 +7,5 @@
     <nav class="my-2 my-md-0 mr-md-3">
         <a class="p-2 text-dark" href="#">Support</a>
     </nav>
-    <a class="btn btn-outline-primary" href="#">Sign in</a>
+    <a class="btn btn-outline-primary" href="/account/login">Sign in</a>
 </div>
